backend/requirements.txt
fastapi
uvicorn
openai
GitPython
python-multipart
typer[all]
requests
rich
PyYAML